Cleanup old implementation
