+++ package.json
@@ -47,7 +47,6 @@
     "react-dom": "^18.2.0",
     "react-plotly.js": "^2.4.0",
     "react-promise-tracker": "^2.1.0",
-    "react-table": "^7.8.0",
     "regenerator-runtime": "^0.13.7",
     "save-svg-as-png": "^1.4.17",
     "shakapacker": "6.5.2",
